% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{auc_maihda}
\alias{auc_maihda}
\title{AUC for logistic MAIHDA (requires pROC)}
\usage{
auc_maihda(object)
}
\arguments{
\item{object}{A "maihda" object fit with binomial family.}
}
\value{
Numeric AUC.
}
\description{
AUC for logistic MAIHDA (requires pROC)
}
