% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maihda_fit.R
\name{maihda_fit}
\alias{maihda_fit}
\title{Fit a MAIHDA model (random intercept for intersectional strata, optional cross-context)}
\usage{
maihda_fit(
  formula,
  data,
  strata,
  family = stats::gaussian(),
  engine = c("lme4", "glmmTMB", "brms"),
  contexts = NULL,
  ...
)
}
\arguments{
\item{formula}{Model formula without random terms (e.g. y ~ 1 or y ~ x1 + x2).}

\item{data}{A data.frame.}

\item{strata}{Character vector that defines the intersectional stratum.}

\item{family}{stats::gaussian() or stats::binomial(), etc.}

\item{engine}{"lme4","glmmTMB","brms".}

\item{contexts}{Optional character vector of additional grouping factors for cross-classified structures,
e.g. c("region") or c("school","area"). Each will get a random intercept \code{(1|context)}.}

\item{...}{Passed to the underlying model function.}
}
\description{
Fit a MAIHDA model (random intercept for intersectional strata, optional cross-context)
}
