% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{maihda_metrics}
\alias{maihda_metrics}
\title{Summarise MAIHDA model metrics}
\usage{
maihda_metrics(object, null = NULL)
}
\arguments{
\item{object}{A "maihda" object.}

\item{null}{Optional "maihda" null model for PCV.}
}
\value{
A one-row tibble with key metrics.
}
\description{
Summarise MAIHDA model metrics
}
