% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pcv}
\alias{pcv}
\title{Proportional Change in Variance (between-stratum)}
\usage{
pcv(null_obj, model_obj)
}
\arguments{
\item{null_obj}{MAIHDA null model (usually y ~ 1 + (1|stratum))}

\item{model_obj}{A richer MAIHDA model (e.g., + main effects)}
}
\description{
Proportional Change in Variance (between-stratum)
}
