% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strata.R
\name{plot_strata}
\alias{plot_strata}
\title{Quick caterpillar plot of stratum effects}
\usage{
plot_strata(strata_df, response = FALSE)
}
\arguments{
\item{strata_df}{Output of predict_strata().}

\item{response}{If TRUE and \code{strata_df} has prob columns, plot probabilities; otherwise plot eta.}
}
\description{
Quick caterpillar plot of stratum effects
}
